+++ src/xlookup.c
@@ -216,19 +216,21 @@ static int xLookupPutAllAsync(XLookupTable *tab, const char *prefix, const XStru
 
     if(f->type == X_STRUCT && recursive) {
       XStructure *sub = (XStructure *) f->value;
+      size_t plen;
       char *p1;
       int count;
 
-      p1 = (char *) malloc(lp + strlen(f->name) + 2 * sizeof(X_SEP) + 12);
+      plen = lp + strlen(f->name) + 2 * sizeof(X_SEP) + 12;
+      p1 = (char *) malloc(plen);
       x_check_alloc(p1);
 
       count = xGetFieldCount(f);
       while(--count >= 0) {
         int n = 0;
-        if(prefix) n = sprintf(p1, "%s" X_SEP, prefix);
+        if(prefix) n = x_snprintf(p1, plen, "%s" X_SEP, prefix);
 
-        if(f->ndim == 0) sprintf(&p1[n], "%s", f->name);
-        else sprintf(&p1[n], "%s" X_SEP "%d", f->name, (count + 1));
+        if(f->ndim == 0) x_snprintf(&p1[n], plen - n, "%s", f->name);
+        else x_snprintf(&p1[n], plen - n, "%s" X_SEP "%d", f->name, (count + 1));
 
         N += xLookupPutAllAsync(tab, p1, &sub[count], TRUE);
       }
@@ -253,19 +255,21 @@ static int xLookupRemoveAllAsync(XLookupTable *tab, const char *prefix, const XS
 
     if(f->type == X_STRUCT && recursive) {
       XStructure *sub = (XStructure *) f->value;
+      size_t plen;
       char *p1;
       int count;
 
-      p1 = (char *) malloc(lp + strlen(f->name) + 2 * sizeof(X_SEP) + 12);
+      plen = lp + strlen(f->name) + 2 * sizeof(X_SEP) + 12;
+      p1 = (char *) malloc(plen);
       x_check_alloc(p1);
 
       count = xGetFieldCount(f);
       while(--count >= 0) {
         int n = 0;
-        if(prefix) n = sprintf(p1, "%s" X_SEP, prefix);
+        if(prefix) n = x_snprintf(p1, plen, "%s" X_SEP, prefix);
 
-        if(f->ndim == 0) sprintf(&p1[n], "%s", f->name);
-        else sprintf(&p1[n], "%s" X_SEP "%d", f->name, (count + 1));
+        if(f->ndim == 0) x_snprintf(&p1[n], plen - n, "%s", f->name);
+        else x_snprintf(&p1[n], plen - n, "%s" X_SEP "%d", f->name, (count + 1));
 
         N += xLookupRemoveAllAsync(tab, p1, &sub[count], TRUE);
       }