+++ CHANGELOG.md
@@ -15,7 +15,7 @@ Upcoming feature release, possibly around 1 August 2026.
 
  - #26: Fixed space allocation in in `xLookupPutAllAsync()` / `xLookupRemoveAllAsync()`.
  
- - #29: `xGetAsDoubleAtIndex()` returned integer rounded values when the field stored `float` or `double` type data. 
+ - #28: `xGetAsDoubleAtIndex()` returned integer rounded values when the field stored `float` or `double` type data. 
    Now, it returns the floating-point value directly.
  
  - Fixed memleak in `xLookupRemove()`.
@@ -49,9 +49,15 @@ Upcoming feature release, possibly around 1 August 2026.
 
 ### Changed
 
+ - #29: Use `snprintf()` instead of `sprintf()` provided it's available. (On older platforms prior to the C99 
+   standard, it defaults to `sprintf()`.)
+
  - Changed `xlookup` hash algorithm to FNV-1a.
  
  - Rounding instead of downcast from double in `xGetAsLongAtIndex()`.
+ 
+ - Use `snprintf()` instead of `sprintf()` wherever possible. For older platforms that do not have `snprintf()`, 
+   define a dummy version, which behaves just like `sprintf()`.
 
  - CMake install to skip `.gitignore` in `examples/`.